+++ semcore/d3-chart/src/types/GroupBar.d.ts
@@ -3,6 +3,7 @@ import { BarContext, BarProps } from './Bar';
 import { HorizontalBarProps } from './HorizontalBar';
 import { Context } from './context';
 import { IntergalacticD3Component } from './Plot';
+import { PatternsConfig } from './Pattern';
 
 /** @deprecated */
 export interface IGroupBarProps extends GroupBarProps, UnknownProperties {}
@@ -14,7 +15,6 @@ export type GroupBarProps = Context & {
   /** Scale for group bars */
   /** @deprecated */
   scaleGroup?: any;
-
   /**
    * The maximum width of each Bar
    */