+++ semcore/tag/src/Tag.jsx
@@ -8,6 +8,7 @@ import { localizedMessages } from './translations/__intergalactic-dynamic-locale
 import i18nEnhance from '@semcore/utils/lib/enhances/i18nEnhance';
 import resolveColorEnhance from '@semcore/utils/lib/enhances/resolveColorEnhance';
 import uniqueIDEnhancement from '@semcore/utils/lib/uniqueID';
+import keyboardFocusEnhance from '@semcore/utils/lib/enhances/keyboardFocusEnhance';
 
 import style from './style/tag.shadow.css';
 import { callAllEventHandlers } from '@semcore/utils/lib/assignProps';
@@ -28,13 +29,21 @@ const legacyThemeRecommendedMigration = {
 class RootTag extends Component {
   static displayName = 'Tag';
   static style = style;
-  static enhance = [i18nEnhance(localizedMessages), uniqueIDEnhancement(), resolveColorEnhance()];
+  static enhance = [
+    i18nEnhance(localizedMessages),
+    uniqueIDEnhancement(),
+    resolveColorEnhance(),
+    keyboardFocusEnhance(),
+  ];
   static defaultProps = {
     theme: 'primary',
     color: 'gray-500',
     size: 'm',
     i18n: localizedMessages,
     locale: 'en',
+  };
+  state = {
+    focusable: 'container',
   };
 
   constructor(props) {
@@ -54,17 +63,45 @@ class RootTag extends Component {
     return { size };
   }
 
-  getCloseProps() {
-    const { getI18nText, id, uid } = this.asProps;
+  getTextProps() {
+    const { interactive } = this.asProps;
+    const id = this.asProps.id || `igc-${this.asProps.uid}-tag`;
+    const { focusable } = this.state;
 
-    return { getI18nText, tagId: id || `igc-${uid}-tag`, uid };
+    return {
+      tabIndex: focusable === 'text' && interactive ? 0 : -1,
+      id: `${id}-text`,
+    };
+  }
+  handleCloseMount = () => {
+    this.setState({ focusable: 'text' });
+  };
+  handleCloseUnmount = () => {
+    this.setState({ focusable: 'container' });
+  };
+  getCloseProps() {
+    const { getI18nText } = this.asProps;
+    const id = this.asProps.id || `igc-${this.asProps.uid}-tag`;
+
+    return {
+      getI18nText,
+      id: `${id}-clear`,
+      'aria-labelledby': `${id}-clear ${id}-text`,
+      'aria-label': getI18nText('remove'),
+      'aria-hidden': 'true',
+      onMount: this.handleCloseMount,
+      onUnmount: this.handleCloseUnmount,
+    };
   }
 
-  handleKeyDown = (e) => {
-    switch (e.key) {
-      case ' ':
+  handleKeyDown = (event) => {
+    switch (event.code) {
+      case 'Space':
       case 'Enter':
-        this.asProps.onClick?.(e);
+        if (this.asProps.onClick) {
+          event.preventDefault();
+          this.asProps.onClick(event);
+        }
         break;
     }
   };
@@ -80,11 +117,11 @@ class RootTag extends Component {
       addonLeft,
       addonRight,
       resolveColor,
-      onClick,
       id: outerId,
       uid,
       onKeyDown,
     } = this.asProps;
+    const { focusable } = this.state;
     const id = outerId || `igc-${uid}-tag`;
 
     return sstyled(styles)(
@@ -92,9 +129,10 @@ class RootTag extends Component {
         render={Box}
         id={id}
         use:interactive={!disabled && interactive}
-        tabIndex={interactive && onClick ? 0 : undefined}
+        role={interactive ? 'button' : undefined}
         tag-color={resolveColor(color)}
         onKeyDown={callAllEventHandlers(onKeyDown, this.handleKeyDown)}
+        use:tabIndex={interactive && focusable === 'container' ? 0 : -1}
       >
         {addonLeft ? <Tag.Addon tag={addonLeft} /> : null}
         {addonTextChildren(Children, Tag.Text, Tag.Addon)}
@@ -109,32 +147,32 @@ function Text(props) {
   const { styles } = props;
   return sstyled(styles)(<SText render={Box} tag='span' />);
 }
+Text.enhance = [keyboardFocusEnhance()];
 
 function Close(props) {
   const SClose = Root;
-  const { styles, getI18nText, tagId, uid } = props;
+  const { styles } = props;
 
-  function onKeyDown(event) {
-    if (props.onKeyDown) {
-      return props.onKeyDown(event);
-    }
+  React.useEffect(() => {
+    props.onMount?.();
+    return () => props.onUnmount?.();
+  }, []);
 
-    if (event.key === 'Enter') {
-      props.onClick?.(event);
-    }
-  }
+  const onKeyDown = React.useCallback(
+    (event) => {
+      if (props.onKeyDown) {
+        return props.onKeyDown(event);
+      }
 
-  return sstyled(styles)(
-    <SClose
-      render={Box}
-      tag={CloseM}
-      interactive
-      id={`igc-${uid}-tag-clear`}
-      aria-labelledby={`igc-${uid}-tag-clear ${tagId}`}
-      aria-label={getI18nText('remove')}
-      onKeyDown={onKeyDown}
-    />,
+      if (props.onClick && (event.code === 'Enter' || event.code === 'Space')) {
+        event.preventDefault();
+        props.onClick(event);
+      }
+    },
+    [props.onKeyDown, props.onClick],
   );
+
+  return sstyled(styles)(<SClose render={Box} tag={CloseM} interactive onKeyDown={onKeyDown} />);
 }
 
 function Addon(props) {