+++ website/docs/data-display/bar-chart/examples/bar.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import { Plot, Bar, YAxis, XAxis } from '@semcore/ui/d3-chart';
+import { scaleLinear, scaleBand } from 'd3-scale';
+
+const Demo = () => {
+  const MARGIN = 40;
+  const width = 500;
+  const height = 300;
+
+  const xScale = scaleBand()
+    .range([MARGIN, width - MARGIN])
+    .domain(data.map((d) => d.category))
+    .paddingOuter(0.2);
+
+  const yScale = scaleLinear()
+    .range([height - MARGIN, MARGIN])
+    .domain([0, 10]);
+
+  return (
+    <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
+      <YAxis>
+        <YAxis.Ticks />
+        <YAxis.Grid />
+      </YAxis>
+      <XAxis>
+        <XAxis.Ticks />
+      </XAxis>
+      <Bar x='category' y='bar' maxBarSize={20} />
+    </Plot>
+  );
+};
+
+const data = Array(5)
+  .fill({})
+  .map((d, i) => ({
+    category: `Category ${i}`,
+    bar: Math.random() * 10,
+  }));
+
+export default Demo;