+++ semcore/card/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.25.0] - 2024-02-14
+
+### Changed
+
+- Version preminor update due to children dependencies update (`@semcore/tooltip` [6.19.0 ~> 6.20.0], `@semcore/icon` [4.25.0 ~> 4.26.0]).
+
+## [5.24.2] - 2024-02-09
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/utils` [4.20.1 ~> 4.20.2], `@semcore/tooltip` [6.18.1 ~> 6.19.0], `@semcore/icon` [4.24.1 ~> 4.25.0]).
+
 ## [5.24.1] - 2024-02-06
 
 ### Changed