+++ semcore/color-picker/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.25.1] - 2024-02-16
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/dropdown` [4.19.2 ~> 4.19.3]).
+
+## [2.25.0] - 2024-02-13
+
+### Changed
+
+- Color palette items are clickable by `Space` now (along with `Enter` as before).
+
+## [2.24.2] - 2024-02-09
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/button` [5.20.1 ~> 5.20.2], `@semcore/icon` [4.24.1 ~> 4.25.0], `@semcore/utils` [4.20.1 ~> 4.20.2]).
+
 ## [2.24.1] - 2024-02-06
 
 ### Changed