+++ semcore/d3-chart/CHANGELOG.md
@@ -2,11 +2,46 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [3.27.2] - 2024-02-07
+## [3.27.5] - 2024-02-21
 
 ### Fixed
 
 - background color for Bar chart from `--intergalactic-chart-palette-order-other-data` to `--intergalactic-chart-grid-bar-chart-base-bg`.
+
+## [3.27.4] - 2024-02-19
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/popper` [5.20.0 ~> 5.20.1]).
+
+## [2.29.0] - 2024-02-16
+
+### Changed
+
+- Removed wrongly added deprecation messages about `ReferenceLine` `value` props.
+- Added `valueEnd` prop to `ReferenceLine.Background` component that allows to specify background width by chart value.
+
+### Fixed
+
+- `patterns` prop was missing in multiple chart types.
+
+## [2.28.0] - 2024-02-13
+
+### Changed
+
+- A11y module links are clickable by `Space` now (along with `Enter` as before).
+
+## [3.27.3] - 2024-02-13
+
+### Fixed
+
+- `Dots` component was missing `value` prop in it's props mapping function.
+
+## [3.27.2] - 2024-02-09
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/animation` [2.18.1 ~> 2.18.2], `@semcore/popper` [5.19.1 ~> 5.20.0], `@semcore/utils` [4.20.1 ~> 4.20.2], `@semcore/core` [2.17.1 ~> 2.17.2]).
 
 ## [3.27.1] - 2024-02-06