+++ semcore/dropdown/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.19.4] - 2024-02-19
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/popper` [5.20.0 ~> 5.20.1]).
+
+## [4.19.3] - 2024-02-16
+
+### Fixed
+
+- Removed deprecation messages about `interaction` property (that were added by mistake).
+
+## [4.19.2] - 2024-02-09
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/popper` [5.19.1 ~> 5.20.0], `@semcore/utils` [4.20.1 ~> 4.20.2], `@semcore/core` [2.17.1 ~> 2.17.2]).
+
 ## [4.19.1] - 2024-02-06
 
 ### Changed