+++ semcore/popper/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.20.1] - 2024-02-16
+
+### Fixed
+
+- Focusing trigger right after popper had been closed by mouse leave was not opening popper again.
+
+## [5.20.0] - 2024-02-06
+
+### Added
+
+- `focusLoop` api that controls how browser focus behaves when goes out of popper.
+
+### Changed
+
+- Invisible focus return and catch elements are removed that were used for keyboard users focus flow control. Instead, trigger may ignore focus event for a short time just after corresponding popper close.
+
 ## [5.19.1] - 2024-02-06
 
 ### Changed