+++ semcore/time-picker/CHANGELOG.md
@@ -2,6 +2,48 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.30.0] - 2024-02-21
+
+### Changed
+
+- Version preminor update due to children dependencies update (`@semcore/select` [4.29.2 ~> 4.30.0]).
+
+## [4.29.2] - 2024-02-19
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/select` [4.29.1 ~> 4.29.2]).
+
+## [4.29.1] - 2024-02-16
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/select` [4.29.0 ~> 4.29.1]).
+
+## [4.29.0] - 2024-02-14
+
+### Changed
+
+- Version preminor update due to children dependencies update (`@semcore/select` [4.28.0 ~> 4.29.0]).
+
+## [4.28.0] - 2024-02-13
+
+### Changed
+
+- Version preminor update due to children dependencies update (`@semcore/select` [4.27.1 ~> 4.28.0]).
+
+## [4.27.1] - 2024-02-09
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/button` [5.20.1 ~> 5.20.2], `@semcore/flex-box` [5.19.0 ~> 5.19.1], `@semcore/input` [4.20.1 ~> 4.20.2], `@semcore/select` [4.27.0 ~> 4.27.1], `@semcore/utils` [4.20.1 ~> 4.20.2], `@semcore/core` [2.17.1 ~> 2.17.2]).
+
+## [4.27.0] - 2024-02-07
+
+### Changed
+
+- Version preminor update due to children dependencies update (`@semcore/select` [4.26.1 ~> 4.27.0]).
+
 ## [4.26.1] - 2024-02-06
 
 ### Changed